factor out polynomial integer root-sweep tests
