+++ cubic/shared/src/test/scala/org/hammerlab/math/polynomial/result/Results.scala
@@ -3,13 +3,13 @@ package org.hammerlab.math.polynomial.result
 import hammerlab.iterator._
 import org.hammerlab.math.polynomial.TestCase
 import org.hammerlab.math.syntax.Doubleish
-import spire.algebra.{ Field, IsReal, NRoot, Trig }
+import spire.algebra.{ Field, IsReal, NRoot, Signed, Trig }
 
 case class Results[D](results: Seq[ResultGroup[D]],
                       errors: Stats)
 
 object Results {
-  def apply[D: Field : Doubleish : IsReal : NRoot : Trig : Solve](cases: Iterator[TestCase[D]]): Results[D] = {
+  def apply[D: Field : Doubleish : NRoot : Signed : Solve](cases: Iterator[TestCase[D]]): Results[D] = {
     val results =
       cases
         .map(Result(_))
