+++ cubic/shared/src/test/scala/org/hammerlab/math/polynomial/test/PolySolverTest.scala
@@ -3,7 +3,6 @@ package org.hammerlab.math.polynomial.test
 import hammerlab.indent.implicits.spaces4
 import hammerlab.iterator._
 import org.hammerlab.math.polynomial.result.Solve
-import hammerlab.lines._
 import hammerlab.math.FromDouble
 import hammerlab.show._
 import org.hammerlab.Suite
@@ -12,6 +11,7 @@ import org.hammerlab.math.format.SigFigs
 import org.hammerlab.math.polynomial
 import org.hammerlab.math.polynomial.result.{ Expected, Results, Stats }
 import org.hammerlab.math.polynomial.roots.RootShapes
+import org.hammerlab.math.polynomial.roots.dsl.IsRootShapes
 import org.hammerlab.math.polynomial.{ ImaginaryRootPair, Real, TestCase, result }
 import org.hammerlab.math.syntax.Doubleish._
 import org.hammerlab.math.syntax.{ Doubleish, E }
@@ -32,9 +32,17 @@ trait ShowD[D] {
     Show { _.toDouble.show }
 }
 
-abstract class PolySolverTest[T : Ordering : FromDouble : Field : IsReal : NRoot : Signed : Trig : Doubleish](degree: Int)
+abstract class PolySolverTest[T : FromDouble : IsReal : NRoot : Signed : Trig](degree: Int)(
+    implicit
+    // name and `val` these bc mix-ins need them
+    val field: Field[T],
+    val ordering: Ordering[T],
+    val doubleish: Doubleish[T]
+)
   extends Suite
-    with ShowD[T] {
+    with ShowD[T]
+    with IntegerRootSweep[T]
+    with IsRootShapes.HasOps {
 
   /** Pretty-printing level */
   implicit var sigfigs: SigFigs = 3
@@ -43,9 +51,6 @@ abstract class PolySolverTest[T : Ordering : FromDouble : Field : IsReal : NRoot
 
   type D = T
 
-  /** Check all polynomials whose roots are comprised of integers in the range [-[[M]], [[M]]], via [[rootSweep]] */
-  val M: Int
-
   val N = degree
 
   /**
@@ -53,68 +58,6 @@ abstract class PolySolverTest[T : Ordering : FromDouble : Field : IsReal : NRoot
    */
   implicit def solve(t: TestCase[D]): Seq[Complex[D]]
   implicit def solver: Solve[D] = solve
-
-  /**
-   * Expected statistics about [[solve]]'s performance against ground-truth on [[TestCase]]s generated by sweeping roots
-   * over integers in the range [-[[M]], [[M]]] (via [[rootSweep]])
-   */
-
-  def sweepTests(M: Int,
-                 expecteds: Map[RootShapes, Stats] = Map.empty): Unit = {
-    val cases = rootSweep(M).toList
-    val n = cases.size
-    cases
-      .groupBy {
-        case TestCase(reals, imags, _, _) ⇒
-          RootShapes(
-            reals.map(_._2),
-            imags.map(_._2)
-          )
-      }
-      .toVector
-      .sortBy(_._1)
-      .foreach {
-        case (shapes, cases) ⇒
-          val expected =
-            expecteds
-              .getOrElse(
-                shapes,
-                throw new Exception(
-                  show"Missing expected stats for shapes: $shapes"
-                )
-              )
-
-          test(show"integer-roots sweep (max $M): root-shapes: ${"%-7s".format(shapes.show)} expected: $expected") {
-            val results = Results(cases.iterator)
-            val actual: Expected = results
-
-            def print(): Unit = {
-              println(
-                show"$shapes:\t$actual"
-              )
-            }
-
-            print()
-
-            def err(e: Exception = null): Unit = {
-              import hammerlab.lines.generic._
-              val lines = actual.errors.lines
-              val msg = lines.show
-              println(indent(lines).show)
-              Option(e)
-                .map(e ⇒ throw new Exception(msg, e))
-                .orElse( throw new Exception(msg   ))
-            }
-
-            try {
-              implicit val ε: E = 1e-2
-              ===(actual.errors, expected)
-            } catch {
-              case e: TestFailedException ⇒ err(e)
-            }
-          }
-      }
-  }
 
   /**
    * Expected statistics about [[solve]]'s performance against ground-truth on [[TestCase]]s with randomly-generated
@@ -227,58 +170,6 @@ abstract class PolySolverTest[T : Ordering : FromDouble : Field : IsReal : NRoot
 
   lazy val numImagPairs = allImaginaryRootPairs.size
 
-  lazy val rootSweepSize =
-    (
-      for {
-        numPairs ← 0 to N/2
-        numReals = N - 2*numPairs
-      } yield
-        numRoots(numPairs, numImagPairs) * numRoots(numReals, 2*M + 1)
-    )
-    .sum
-
-  /**
-   * Iterate over the available imaginary-root conjugate-pairs (with integral real and imaginary parts in [-M, M], per
-   * [[allImaginaryRootPairs]]), emitting all possible multi-sets of size `num`, and allowing repeats of individual
-   * pairs (representing multiple roots)
-   */
-  def imaginaryRootPairs(num: Int) =
-    allImaginaryRootPairs
-      .unorderedSubsetsWithReplacement(num)
-
-  /**
-   * Iterator over all sets of `num` integer-roots in the range [-M,M] (including double/triple/quadruple roots)
-   */
-  def realRootsIter(num: Int): Iterator[List[(Real, Int)]] =
-    (-M to M)
-      .unorderedSubsetsWithReplacement(num)
-      .map {
-        _.map {
-          case (value, arity) ⇒
-            Real(value: D) → arity
-        }
-      }
-
-  /**
-   * Generate roots and corresponding coefficients for all possible mixtures of real/imaginary roots with coefficients
-   * in [-M,m]
-   */
-  def rootSweep(M: Int = M): Iterator[TestCase[D]] =
-    for {
-      numImaginaryRootPairs ← (0 to N/2).iterator
-      numImaginaryRoots = 2 * numImaginaryRootPairs
-
-      numRealRoots = N - numImaginaryRoots
-      reals ← realRootsIter(numRealRoots)
-
-      imags ← imaginaryRootPairs(numImaginaryRootPairs)
-    } yield
-      TestCase[D](
-        reals,
-        imags,
-        1
-      )
-
   def gaussian(): D = nextGaussian()
 
   def logNormal(): D =
@@ -349,4 +240,8 @@ abstract class PolySolverTest[T : Ordering : FromDouble : Field : IsReal : NRoot
         )
     )
   }
+
+  implicit def convTuple[L, R](t: (L, R))(implicit f: IsRootShapes[L]): (RootShapes, R) = (f(t._1), t._2)
+
+  implicit def makeResults(cases: Iterator[TestCase[D]]): Results = Results(cases)
 }
